+++ crates/backend/src/consts.rs
@@ -1,0 +1,5 @@
+pub const DEFAULT_SCHEMA: &str = indoc::indoc! {"
+    type Placeholder @model {
+        id: ID!
+    }
+"};