+++ crates/cli/src/cli_input.rs
@@ -12,6 +12,7 @@ pub fn build_cli() -> Command<'static> {
             Command::new("dev").about("Run your grafbase project locally").args(&[
                 arg!(-p --port <port> "Use a specific port")
                     .takes_value(true)
+                    .default_value("4000")
                     .value_parser(value_parser!(u16))
                     .required(false),
                 arg!(-s --search "If a given port is unavailable, search for another"),
@@ -28,15 +29,17 @@ pub fn build_cli() -> Command<'static> {
                     To use, write the output to the appropriate location for your shell
                 "}),
         )
-    // .subcommand(Command::new("login").about("TBD"))
     // .subcommand(
-    //     Command::new("create")
+    //     Command::new("init")
     //         .args(&[
-    //             arg!(<name> "the name of the project to create"),
+    //             arg!([name] "the name of the project to create"),
     //             arg!(-t --template "the name of the template to use for the new project"),
     //         ])
-    //         .about("TBD"),
+    //         .about(indoc! {"
+    //             Creates a grafbase/schema.graphql file and folder in the current directory
+    //         "}),
     // )
+    // .subcommand(Command::new("login").about("TBD"))
     // .subcommand(Command::new("deploy").about("TBD"))
     // .subcommand(Command::new("logs").about("TBD"))
     // // TODO: schema edit / view