+++ crates/cli/src/panic_hook/mod.rs
@@ -1,9 +1,12 @@
-mod report;
+pub mod report;
 
 use report::{Method, Report};
 use std::borrow::Cow;
 use std::panic::PanicInfo;
 use std::path::{Path, PathBuf};
+use watercolor::watercolor;
+
+use crate::watercolor;
 
 /// A convenient metadata struct that describes a crate
 pub struct Metadata {
@@ -13,31 +16,38 @@ pub struct Metadata {
     pub name: Cow<'static, str>,
     /// The URL of the crate's website
     pub homepage: Cow<'static, str>,
+    /// The URL of the crate's repository
+    pub repository: Cow<'static, str>,
+    /// A discord invite link
+    pub discord: Cow<'static, str>,
 }
 
 /// `panic-hook` initialisation macro
 ///
 /// ```
-/// panic_hook::setup!();
+/// panic_hook!();
 /// ```
 #[macro_export]
-macro_rules! setup {
+macro_rules! panic_hook {
     () => {
         #[allow(unused_imports)]
         use std::panic::{self, PanicInfo};
         #[allow(unused_imports)]
-        use $crate::{handle_dump, print_msg, Metadata};
+        use $crate::panic_hook::{handle_dump, print_msg, Metadata};
 
+        #[allow(unused_variables)]
         let meta = Metadata {
             version: env!("CARGO_PKG_VERSION").into(),
-            name: "Grafbase CLI".into(),
+            name: "the Grafbase CLI".into(),
             homepage: env!("CARGO_PKG_HOMEPAGE").into(),
+            repository: "https://github.com/grafbase/grafbase".into(),
+            discord: "https://discord.gg/grafbase".into(),
         };
 
         #[cfg(not(debug_assertions))]
         match ::std::env::var("RUST_BACKTRACE") {
             Err(_) => {
-                panic::set_hook(Box::new(move |info: &PanicInfo| {
+                panic::set_hook(Box::new(move |info: &PanicInfo<'_>| {
                     let file_path = handle_dump(&meta, info);
                     print_msg(file_path, &meta);
                     process::exit(exitcode::SOFTWARE);
@@ -49,33 +59,45 @@ macro_rules! setup {
 }
 
 /// Utility function that prints a message to our human users
+#[allow(dead_code)]
 pub fn print_msg<P: AsRef<Path>>(file_path: Option<P>, metadata: &Metadata) {
-    let Metadata { name, homepage, .. } = metadata;
+    let Metadata {
+        name,
+        homepage,
+        repository,
+        discord,
+        ..
+    } = metadata;
     let file_path = match file_path {
         Some(file_path) => format!("{}", file_path.as_ref().display()),
         None => "<Failed to store file to disk>".to_string(),
     };
 
-    eprintln!(
-        indoc::indoc! {r##"
-            Well, this is embarrassing.
-            {name} had a problem and crashed.
+    watercolor::output_error!(
+        indoc::indoc!{r##"
+            Well, this is embarrassing, {} had a problem and crashed.
             
-            We have generated a report file at "{file_path}". 
-            To help us address this issue, please consider submitting a GitHub issue or sending us an email with the subject of "{name} Crash Report" and include the report as an attachment.
-
-            - Homepage: {homepage}
-
+            We have generated a report file at {}. 
+            To help us address this issue, please consider submitting a GitHub issue or sending a message on Discord and including the report.
+        
+            - Homepage: {}
+            - Repository: {}
+            - Discord: {}
+        
             Thank you!
         "##},
-        name = name,
-        file_path = file_path,
-        homepage = homepage
+        name,
+        watercolor!("{file_path}", @BrightBlue),
+        watercolor!("{homepage}", @BrightBlue),
+        watercolor!("{repository}", @BrightBlue),
+        watercolor!("{discord}", @BrightBlue),
+        @BrightRed
     );
 }
 
 /// Utility function which will handle dumping information to disk
 #[must_use]
+#[allow(dead_code)]
 pub fn handle_dump(meta: &Metadata, panic_info: &PanicInfo<'_>) -> Option<PathBuf> {
     let mut explanation = String::new();