+++ crates/dev-server/src/types.rs
@@ -1,7 +1,7 @@
 use rust_embed::RustEmbed;
 
 #[derive(RustEmbed)]
-#[folder = "../../assets/"]
+#[folder = "assets/"]
 pub struct Assets;
 
 #[derive(Clone, Copy)]